add flag arguments

+++ ani2xcur.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -8,47 +9,55 @@ import (
 )
 
 var TEMPDIR string
-var OUTPUTDIR string = "."
+var OUTPUTDIR string
+
+func init() {
+	flag.Usage = func() {
+		fmt.Fprintf(flag.CommandLine.Output(),
+			"usage: %s [options] files...\n\nconverts windows cursor files to to xcur files\n\n", os.Args[0])
+		flag.PrintDefaults()
+	}
+	flag.StringVar(&OUTPUTDIR, "o", ".", "output `dir`ectory")
+	flag.StringVar(&TEMPDIR, "tmp", os.TempDir(), "temporary `dir`ectory [for debugging]")
+}
+
+func ensureDirectory(name string) {
+	finfo, err := os.Stat(name)
+	if os.IsNotExist(err) {
+		if err = os.Mkdir(name, 0755); err != nil {
+			fmt.Fprintf(os.Stderr, "error when creating %v: %v\n", name, err)
+			os.Exit(2)
+		}
+	} else if !finfo.IsDir() {
+		fmt.Fprintln(os.Stderr, "error:", name, "is not a directory")
+		os.Exit(20)
+	}
+}
 
 func main() {
-	var (
-		input []string
-		err   error
-	)
-	switch len(os.Args) {
-	case 0:
-		panic("unreachable")
-	case 1:
-		fmt.Println("usage: ani2xcur input-file [input-file]... [output-directory]")
-		os.Exit(22)
-	case 2:
-		input = []string{os.Args[1]}
-	default:
-		l := len(os.Args)
-		input = os.Args[1 : l-1]
-		OUTPUTDIR = os.Args[l-1]
+	var err error
+	flag.Parse()
 
-		finfo, err := os.Stat(OUTPUTDIR)
-		if os.IsNotExist(err) {
-			if err := os.Mkdir(OUTPUTDIR, 0755); err != nil {
-				fmt.Fprintf(os.Stderr, "error when creating %v: %v\n", OUTPUTDIR, err)
-				os.Exit(2)
-			}
-		} else if !finfo.IsDir() {
-			fmt.Fprintln(os.Stderr, "error:", OUTPUTDIR, "is not a directory")
-			os.Exit(20)
+	ensureDirectory(OUTPUTDIR)
+
+	if TEMPDIR == os.TempDir() {
+		TEMPDIR, err = ioutil.TempDir(os.TempDir(), "ani2xcur-")
+		defer os.RemoveAll(TEMPDIR)
+		defer os.Remove(TEMPDIR)
+		if err != nil {
+			fmt.Fprintln(os.Stderr, "error making temporary directory:", err)
+			os.Exit(1)
 		}
+	} else {
+		ensureDirectory(TEMPDIR)
 	}
 
-	TEMPDIR, err = ioutil.TempDir(os.TempDir(), "ani2xcur-")
-	defer os.RemoveAll(TEMPDIR)
-	defer os.Remove(TEMPDIR)
-	if err != nil {
-		fmt.Fprintln(os.Stderr, "error making temporary directory:", err)
+	if len(flag.Args()) == 0 {
+		flag.Usage()
 		os.Exit(1)
 	}
 
-	for _, f := range input {
+	for _, f := range flag.Args() {
 		ext := filepath.Ext(f)
 		switch ext {
 		case ".ico":
